flux-content: move spam to test util
Problem: flux-content spam is just a test program that shouldn't
be in a user facing program.

Move subcommand to a standalone test program.
Update sharness tests that use it.
